refactor: extract field cache & schema to lib/sp/fieldCache.ts (Phase 2)
