+++ src/lib/spClient.ts
@@ -144,9 +144,7 @@ const DEFAULT_LIST_TEMPLATE = 100;
 
 // ─── Types & schemas imported from @/lib/sp/types (SSOT) ────────────────────
 import {
-    escapeXml,
     trimGuidBraces,
-    withGuidBraces,
     type JsonRecord,
     type RetryReason
 } from '@/lib/sp/types';
@@ -180,85 +178,20 @@ import type {
 } from '@/lib/sp/types';
 
 
-// ────────────────────────────────────────────────────────────────────────────
-// Fields Cache（sessionStorage）
-// ────────────────────────────────────────────────────────────────────────────
-const FIELDS_CACHE_TTL_MS = 20 * 60 * 1000; // 20分
-
-
-
-function nowMs(): number {
-  return Date.now();
-}
-
-function safeJsonParse<T>(raw: string | null): T | null {
-  if (!raw) return null;
-  try {
-    return JSON.parse(raw) as T;
-  } catch {
-    return null;
-  }
-}
-
-function safeJsonStringify(obj: unknown): string | null {
-  try {
-    return JSON.stringify(obj);
-  } catch {
-    return null;
-  }
-}
-
-function makeFieldsCacheKey(siteUrl: string, listTitle: string): string {
-  return `sp.fieldsCache.v1::${siteUrl}::${listTitle}`;
-}
-
-
-
-const buildFieldSchema = (def: SpFieldDef): string => {
-  const attributes: string[] = [];
-  const addAttr = (key: string, raw: string | number | boolean | undefined) => {
-    if (raw === undefined || raw === null || raw === '') return;
-    const value = typeof raw === 'boolean' ? (raw ? 'TRUE' : 'FALSE') : String(raw);
-    attributes.push(`${key}="${escapeXml(value)}"`);
-  };
-
-  addAttr('Name', def.internalName);
-  addAttr('StaticName', def.internalName);
-  addAttr('DisplayName', def.displayName ?? def.internalName);
-  addAttr('Type', def.type);
-  if (def.required) addAttr('Required', 'TRUE');
-  if (def.richText) addAttr('RichText', 'TRUE');
-  if (def.dateTimeFormat) addAttr('Format', def.dateTimeFormat);
-  if (def.type === 'Lookup') {
-    if (def.lookupListId) addAttr('List', withGuidBraces(def.lookupListId));
-    addAttr('ShowField', def.lookupFieldName ?? 'Title');
-    if (def.allowMultiple) addAttr('Mult', 'TRUE');
-  } else if (def.allowMultiple) {
-    addAttr('Mult', 'TRUE');
-  }
-
-  if (def.type === 'Boolean' && typeof def.default === 'boolean') {
-    addAttr('Default', def.default ? '1' : '0');
-  }
-
-  const inner: string[] = [];
-  if (def.description) {
-    inner.push(`<Description>${escapeXml(def.description)}</Description>`);
-  }
-  if ((def.type === 'Choice' || def.type === 'MultiChoice') && def.choices?.length) {
-    const choiceXml = def.choices.map((choice) => `<CHOICE>${escapeXml(choice)}</CHOICE>`).join('');
-    inner.push(`<CHOICES>${choiceXml}</CHOICES>`);
-    if (def.default && typeof def.default === 'string') {
-      inner.push(`<Default>${escapeXml(def.default)}</Default>`);
-    }
-  } else if (def.default !== undefined && def.type !== 'Boolean') {
-    inner.push(`<Default>${escapeXml(String(def.default))}</Default>`);
-  }
-
-  const attrs = attributes.join(' ');
-  const body = inner.join('');
-  return `<Field ${attrs}>${body}</Field>`;
-};
+// ─── Field cache & schema imported from @/lib/sp/fieldCache (SSOT) ──────────
+import {
+    buildFieldSchema,
+    buildSelectFields,
+    extractMissingField,
+    FIELDS_CACHE_TTL_MS,
+    getMissingSet,
+    makeFieldsCacheKey,
+    markOptionalMissing,
+    nowMs,
+    resetMissingOptionalFieldsCache,
+    safeJsonParse,
+    safeJsonStringify,
+} from '@/lib/sp/fieldCache';
 
 const sanitizeEnvValue = (value: unknown): string => (typeof value === 'string' ? value.trim() : '');
 
@@ -271,38 +204,8 @@ const USERS_OPTIONAL_FIELDS = ['FullNameKana', 'Furigana', 'Email', 'Phone', 'Bi
 const STAFF_BASE_FIELDS = ['Id', 'StaffID', 'StaffName', 'Role', 'Phone', 'Email'] as const;
 const STAFF_OPTIONAL_FIELDS = ['StaffID', 'AttendanceDays', 'Certifications', 'Department', 'Notes'] as const;
 
-const missingOptionalFieldsCache = new Map<string, Set<string>>();
-
-const getMissingSet = (listTitle: string): Set<string> => {
-  let current = missingOptionalFieldsCache.get(listTitle);
-  if (!current) {
-    current = new Set<string>();
-    missingOptionalFieldsCache.set(listTitle, current);
-  }
-  return current;
-};
-
-const markOptionalMissing = (listTitle: string, field: string) => {
-  if (!field) return;
-  getMissingSet(listTitle).add(field);
-};
-
-const extractMissingField = (message: string): string | null => {
-  const match = message.match(/'([^']+)'/);
-  if (match && match[1]) {
-    return match[1];
-  }
-  return null;
-};
-
-const buildSelectFields = (baseFields: readonly string[], optionalFields: readonly string[], missing: Set<string>): string[] => {
-  const base = baseFields.filter((field) => !missing.has(field));
-  const optional = optionalFields.filter((field) => !missing.has(field));
-  const merged = [...base, ...optional];
-  return Array.from(new Set(merged));
-};
-
 const normalizeGuidCandidate = (value: string): string => trimGuidBraces(value.replace(/^guid:/i, ''));
+
 
 const buildListItemsPath = (listTitle: string, select: string[], top: number): string => {
   const queryParts: string[] = [];
@@ -1467,40 +1370,13 @@ export async function createSchedule<T extends Record<string, unknown>>(_sp: Use
 
 export const __ensureListInternals = { buildFieldSchema };
 
-/**
- * Fields キャッシュを手動クリア（デバッグ用）
- */
-export function clearFieldsCacheFor(listTitle: string, siteUrl?: string): void {
-  if (typeof sessionStorage === 'undefined') return;
-  const url = siteUrl || ensureConfig().baseUrl;
-  const key = makeFieldsCacheKey(url, listTitle);
-  sessionStorage.removeItem(key);
-  console.log('[spClient][fieldsCache] 🗑️ cleared', { listTitle });
-}
-
-/**
- * 全 Fields キャッシュをクリア
- */
-export function clearAllFieldsCache(): void {
-  if (typeof sessionStorage === 'undefined') return;
-  const prefix = 'sp.fieldsCache.v1::';
-  let count = 0;
-  for (let i = sessionStorage.length - 1; i >= 0; i--) {
-    const key = sessionStorage.key(i);
-    if (key?.startsWith(prefix)) {
-      sessionStorage.removeItem(key);
-      count++;
-    }
-  }
-  console.log('[spClient][fieldsCache] 🗑️ cleared all', { count });
-}
+// Re-export cache clear utilities from fieldCache
+export { clearAllFieldsCache, clearFieldsCacheFor } from '@/lib/sp/fieldCache';
 
 // test-only export (intentionally non-exported in production bundles usage scope)
 export const __test__ = {
   ensureConfig,
-  resetMissingOptionalFieldsCache(): void {
-    missingOptionalFieldsCache.clear();
-  },
+  resetMissingOptionalFieldsCache,
   resolveStaffListIdentifier,
 };
